Progress towards finding notes in the chord differently
I changed the way chord notes are computed. Now it is based on the root and quality, so it is easy to include non diatonic chords as well as species 7th chords. The tonality classes are partially updated (may require further changes) and the constraints to find the notes and the bass are already modified (behaviour doesn't change, only implementation). The other constraints must also be adapted to work with this new implementation; for now some don't work anymore, this is why the FourVoiceTexture.cpp class has not been updated in this commit. It will be commited once it is working again.